+++ src/app/components/filesystem-browser/table-cell-json-info.js
@@ -112,26 +112,39 @@ export default Component.extend(I18n, {
 
   async resolveSelectedJsonString() {
     const jsonObj = this.json;
+    const result = {
+      isValid: true,
+      content: '',
+    };
     if (!this.hasJsonMetadata || jsonObj === undefined) {
-      return '';
+      return result;
     }
     if (this.queryType === 'all') {
-      return this.stringifyJson(jsonObj);
+      result.content = this.stringifyJson(jsonObj);
+      return result;
     }
     if (this.queryType === 'key') {
       if (!this.jsonKey || !jsonObj) {
-        return '';
+        return result;
       }
       const selectedData = jsonObj[this.jsonKey];
-      return this.stringifyJson(selectedData);
+      result.content = this.stringifyJson(selectedData);
+      return result;
     }
     if (this.queryType === 'query') {
       if (!this.jsonQuery || !jsonObj) {
-        return '';
+        return result;
       }
-      const expr = jsonata(this.jsonQuery);
-      const result = await expr.evaluate(jsonObj);
-      return this.stringifyJson(result);
+      try {
+        const expr = jsonata(this.jsonQuery);
+        const selectedData = await expr.evaluate(jsonObj);
+        result.content = this.stringifyJson(selectedData);
+        return result;
+      } catch (error) {
+        result.isValid = false;
+        result.content = error.message;
+        return result;
+      }
     }
   },
 
@@ -145,7 +158,7 @@ export default Component.extend(I18n, {
     'tooltipMaxLines',
     'maxTooltipLineLength',
     function tooltipSpec() {
-      const lines = this.selectedJsonString.split('\n');
+      const lines = this.selectedJsonString.content.split('\n');
 
       let jsonText = '';
       let currentLineNumber = 0;
@@ -198,13 +211,13 @@ export default Component.extend(I18n, {
    * @type {ComputedProperty<{ lines: Array<string>, isTruncated: boolean }>}
    */
   htmlContentSpec: computed('selectedJsonString', function htmlContentSpec() {
-    if (!this.selectedJsonString) {
+    if (!this.selectedJsonString?.isValid || !this.selectedJsonString?.content) {
       return {
         lines: [],
         isTruncated: false,
       };
     }
-    const preFormattedText = this.selectedJsonString
+    const preFormattedText = this.selectedJsonString.content
       .replace(/(\r\n|\n|\r)/gm, '')
       .replace(/</g, '&lt;')
       .replace(/>/g, '&gt;');
@@ -343,7 +356,7 @@ export default Component.extend(I18n, {
     },
     copyJson(event) {
       event.stopPropagation();
-      this.globalClipboard.copy(this.selectedJsonString);
+      this.globalClipboard.copy(this.selectedJsonString?.content);
     },
   },
 });